Revert "Use floating point widths in KviIrcView: fixes overlapping texts"
This reverts commit 148d8d9e1b6530676093f660ec64f6a2ee673235.

+++ src/kvirc/ui/KviIrcView_private.h
@@ -102,7 +102,7 @@ struct KviIrcViewWrappedBlock
 	KviIrcViewLineChunk * pChunk; // pointer to real line chunk or 0 for word wraps
 	int block_start;              // this is generally different than pAttribute->block_idx!
 	int block_len;                // length if the block in characters
-	float block_width;              // width of the block in pixels
+	int block_width;              // width of the block in pixels
 } _KVI_PACKED;
 
 struct KviIrcViewLine
@@ -135,11 +135,11 @@ struct KviIrcViewWrappedBlockSelectionInfo
 {
 	int selection_type;
 	int part_1_length;
-	float part_1_width;
+	int part_1_width;
 	int part_2_length;
-	float part_2_width;
+	int part_2_width;
 	int part_3_length;
-	float part_3_width;
+	int part_3_width;
 };
 
 #ifdef COMPILE_ON_WINDOWS
